sdh - game manager update, minor changes to players to allow screen clearing
stop button now halts the game and sets flag to display a blank screen.
this allows switching between 1 and 2 player games without residual
display of previous game (it's there, just not visible). stop acts like
as pause if the game is not over.

Added boolean clearScreen to player classes so paintComponent function
knows when to display a blank screen for that game.  ball and paddle
movement continues if "playing" AND not gameOver. This allows the "game
over" screen to remain until the game is stopped.

in 2 player, added reset function (as in 1-player)and fixed scoring to
aid in testing.

oh and for testing , I lowered the winning score to 4. we'll want to put
that back.  also, we had talked about unlimited  hits  within a time
period.